pc-6 add a new file for testing navigation by categories search, wip
